Fix ci (#430)
* fix ci

* remove designated initializers to prevent CI failures

+++ simpleble/test/src/test_bytearray.cpp
@@ -112,32 +112,32 @@ TEST(ByteArrayTest, DefaultConstructor_Empty) {
 
 TEST(ByteArrayTest, PushBackIncreasesSize) {
     bytearray byteArray;
-    byteArray.push_back(0x01);
+    byteArray.push_back(uint8_t(0x01));
     EXPECT_EQ(1, byteArray.size());
-    byteArray.push_back(0x02);
+    byteArray.push_back(uint8_t(0x02));
     EXPECT_EQ(2, byteArray.size());
 }
 
 TEST(ByteArrayTest, ClearEmptiesArray) {
     bytearray byteArray;
-    byteArray.push_back(0x01);
-    byteArray.push_back(0x02);
+    byteArray.push_back(uint8_t(0x01));
+    byteArray.push_back(uint8_t(0x02));
     byteArray.clear();
     EXPECT_TRUE(byteArray.empty());
 }
 
 TEST(ByteArrayTest, IndexOperatorAccessesCorrectElement) {
     bytearray byteArray;
-    byteArray.push_back(0x01);
-    byteArray.push_back(0x02);
+    byteArray.push_back(uint8_t(0x01));
+    byteArray.push_back(uint8_t(0x02));
     EXPECT_EQ(0x01, byteArray[0]);
     EXPECT_EQ(0x02, byteArray[1]);
 }
 
 TEST(ByteArrayTest, DataPointerIsValid) {
     bytearray byteArray;
-    byteArray.push_back(0x01);
-    byteArray.push_back(0x02);
+    byteArray.push_back(uint8_t(0x01));
+    byteArray.push_back(uint8_t(0x02));
     const uint8_t* data_ptr = byteArray.data();
     ASSERT_NE(nullptr, data_ptr);
     EXPECT_EQ(0x01, data_ptr[0]);
@@ -146,8 +146,8 @@ TEST(ByteArrayTest, DataPointerIsValid) {
 
 TEST(ByteArrayTest, BeginEndIterators) {
     bytearray byteArray;
-    byteArray.push_back(0x01);
-    byteArray.push_back(0x02);
+    byteArray.push_back(uint8_t(0x01));
+    byteArray.push_back(uint8_t(0x02));
     auto it = byteArray.begin();
     EXPECT_EQ(0x01, *it);
     ++it;
@@ -158,8 +158,8 @@ TEST(ByteArrayTest, BeginEndIterators) {
 
 TEST(ByteArrayTest, ConstIndexOperatorAccessesCorrectElement) {
     bytearray byteArray;
-    byteArray.push_back(0x01);
-    byteArray.push_back(0x02);
+    byteArray.push_back(uint8_t(0x01));
+    byteArray.push_back(uint8_t(0x02));
     const bytearray constByteArray = byteArray;
     EXPECT_EQ(0x01, constByteArray[0]);
     EXPECT_EQ(0x02, constByteArray[1]);
